Added backend code for drink items and for bartender!

+++ src/main/java/akatsuki/restaurantsysteminformation/unregistereduser/UnregisteredUserService.java
@@ -17,5 +17,5 @@ public interface UnregisteredUserService {
 
     List<UnregisteredUser> getAll();
 
-    UnregisteredUser checkPinCode(long pinCode, UserType type);
+    UnregisteredUser checkPinCode(int pinCode, UserType type);
 }
